rename reflection dir to cale_info
