+++ effects/extrinsic.ts
@@ -2,8 +2,8 @@ import { unimplemented } from "../deps/std/testing/asserts.ts"
 import * as Z from "../deps/zones.ts"
 import { Signer } from "../frame_metadata/mod.ts"
 import { MultiAddress } from "../primitives/mod.ts"
-import { era } from "../reflection/mod.ts"
 import * as rpc from "../rpc/mod.ts"
+import { era } from "../scale_info/mod.ts"
 import * as U from "../util/mod.ts"
 import { const as const_ } from "./const.ts"
 import { metadata } from "./metadata.ts"
