+++ frame_metadata/Key.ts
@@ -1,5 +1,5 @@
 import * as $ from "../deps/scale.ts"
-import { DeriveCodec } from "../reflection/Codec.ts"
+import { DeriveCodec } from "../scale_info/Codec.ts"
 import { hashers } from "../util/mod.ts"
 import { HasherKind, MapStorageEntryType, Pallet, StorageEntry } from "./Metadata.ts"
