+++ frame_metadata/Metadata.ts
@@ -1,5 +1,5 @@
 import * as $ from "../deps/scale.ts"
-import { $tyId, $tys, Ty } from "../reflection/Ty.ts"
+import { $tyId, $tys, Ty } from "../scale_info/Ty.ts"
 import * as U from "../util/mod.ts"
 
 export type HasherKind = $.Native<typeof $hasherKind>
