+++ mod.ts
@@ -6,6 +6,6 @@ export * as fluent from "./fluent/mod.ts"
 export * as M from "./frame_metadata/mod.ts"
 export { type Signer } from "./frame_metadata/mod.ts"
 export * from "./primitives/mod.ts"
-export { $null, ChainError } from "./reflection/mod.ts"
 export * as rpc from "./rpc/mod.ts"
+export { $null, ChainError } from "./scale_info/mod.ts"
 export { End, getOrInit, type Hex, hex, type Listener, ss58, throwIfError } from "./util/mod.ts"
