+++ scale_info/Codec.ts
