+++ scale_info/TyVisitor.ts
