+++ scale_info/redirects/Era.ts
